refactor trainer to prevent circular dependencies later
fix loader default
